requirements.txt
Django==3.0.8
django-celery==3.3.1
django-credit-cards==0.4.1
django-crispy-forms==1.11.0
django-crontab===0.7.1
django-heroku==0.3.1
django-currentuser==0.5.1
django-filter==2.4.0
django-multiselectfield==0.1.12
django-otp==0.9.3
django-pandas==0.6.2
django-pure-pagination==0.3.0
django-rest-framework==0.1.0
django-slick-reporting==0.5.5
django-tinymce==3.2.0
djangorestframework==3.12.1
gunicorn==20.0.4
html5lib==1.1
numpy==1.19.1
pandas==1.1.2
python-dateutil==2.8.1
requests==2.22.0
simplejson==3.17.2
urllib3==1.25.6
waitress==1.4.4
APScheduler==3.0.0
whitenoise==5.2.0

django-common-helpers==0.9.2